+++ starters/gatsby-starter-catalyst-bery/sanity-studio/schemas/schema.js
@@ -20,6 +20,8 @@ import theme from "./theme"
 import themeColors from "./themeColors"
 import footerColors from "./footerColors"
 import headerColors from "./headerColors"
+import siteHeader from "./siteHeader"
+import category from "./category"
 
 // Then we give our schema to the builder and provide the result to Sanity
 export default createSchema({
@@ -45,5 +47,7 @@ export default createSchema({
     themeColors,
     headerColors,
     footerColors,
+    siteHeader,
+    category,
   ]),
 })