+++ starters/gatsby-starter-catalyst-bery/sanity-studio/structure/deskStructure.js
@@ -5,8 +5,8 @@ import {
   MdList,
   MdInfo,
   MdPerson,
-  MdDashboard,
 } from "react-icons/lib/md"
+import { RiLayoutTopLine } from "react-icons/ri"
 
 const hiddenDocTypes = (listItem) =>
   ![
@@ -20,6 +20,8 @@ const hiddenDocTypes = (listItem) =>
     "theme",
     "themeLmColor",
     "themeDmColor",
+    "siteHeader",
+    "category",
   ].includes(listItem.getId())
 
 export default () =>
@@ -56,27 +58,32 @@ export default () =>
                 .title("Social Links")
                 .icon(MdShare)
                 .child(S.documentTypeList("socialLink").title("Social Links")),
-              S.listItem()
-                .title("Theme-UI")
-                .icon(MdDashboard)
-                .child(
-                  S.document()
-                    .schemaType("theme")
-                    .documentId("theme")
-                    .title("Theme-UI Theme")
-                ),
             ])
         ),
       // Add a visual divider (optional)
       S.divider(),
       S.listItem()
+        .title("Header")
+        .icon(RiLayoutTopLine)
+        .child(S.editor().schemaType("siteHeader").documentId("siteHeader")),
+      S.listItem()
         .title("Pages")
         .child(S.documentTypeList("page").title("Pages")),
       S.listItem()
         .title("Posts")
-        .child(S.documentTypeList("post").title("Posts")),
-      S.listItem()
-        .title("Projects")
-        .child(S.documentTypeList("project").title("Projects")),
+        .child(
+          S.list()
+            .title("Posts")
+            .items([
+              S.listItem()
+                .title("All Posts")
+                .schemaType("post")
+                .child(S.documentTypeList("post").title("All Posts")),
+              S.listItem()
+                .title("Categories")
+                .schemaType("category")
+                .child(S.documentTypeList("category").title("Categories")),
+            ])
+        ),
       ...S.documentTypeListItems().filter(hiddenDocTypes),
     ])