+++ themes/gatsby-theme-catalyst-bery/src/components/post-list-template.js
@@ -25,70 +25,77 @@ const PostsTemplate = ({ data }) => {
         {sanityDisplayPostListTitle && (
           <Styled.h1>{sanityPostListTitle}</Styled.h1>
         )}
-        {posts.map((post) => (
-          <article
-            sx={{
-              mb: 5,
-            }}
-            key={post.id}
-          >
-            <Styled.a
-              as={Link}
-              to={rootPath.concat(post.slug.current.replace(/\/*$/, `/`))}
-            >
-              <Img
-                fluid={post.featuredImage.asset.fluid}
-                sx={{
-                  height: "250px",
-                  borderRadius: "5px",
-                }}
-              />
-            </Styled.a>
-            <Styled.a
-              as={Link}
-              to={rootPath.concat(post.slug.current.replace(/\/*$/, `/`))}
-              style={{ textDecoration: "none" }}
-            >
-              <Styled.h2
-                sx={{
-                  mt: 2,
+        <div sx={{ my: 5 }}>
+          {posts.map((post) => (
+            <article
+              sx={{
+                mb: 5,
+                ":last-of-type": {
                   mb: 0,
-                  fontSize: [4, null, null, 5, null],
+                },
+              }}
+              key={post.id}
+            >
+              <Styled.a
+                as={Link}
+                to={rootPath.concat(post.slug.current.replace(/\/*$/, `/`))}
+              >
+                <Img
+                  fluid={post.featuredImage.asset.fluid}
+                  sx={{
+                    height: "250px",
+                    borderRadius: "5px",
+                  }}
+                />
+              </Styled.a>
+              <Styled.a
+                as={Link}
+                to={rootPath.concat(post.slug.current.replace(/\/*$/, `/`))}
+                style={{ textDecoration: "none" }}
+              >
+                <Styled.h2
+                  sx={{
+                    mt: 2,
+                    mb: 0,
+                    color: "primary",
+                    fontSize: [4, null, null, 5, null],
+                    transition: "color 0.2s ease",
+                    ":hover, :focus, :active": {
+                      color: "secondary",
+                    },
+                  }}
+                >
+                  {post.title}
+                </Styled.h2>
+              </Styled.a>
+              <Styled.p sx={{ fontSize: 1, color: "textGray", m: 0, mt: 2 }}>
+                {post.date} &bull;{" "}
+                <FaRegClock
+                  sx={{
+                    position: "relative",
+                    top: "0.125em",
+                  }}
+                />{" "}
+                {post.readingTimeInMinutes} Min
+              </Styled.p>
+              <Styled.p sx={{ mt: 2, mb: 0 }}>{post.excerpt}</Styled.p>
+              <Styled.a
+                as={Link}
+                to={rootPath.concat(post.slug.current.replace(/\/*$/, `/`))}
+                sx={{
+                  color: "primary",
                   transition: "color 0.2s ease",
                   ":hover, :focus, :active": {
                     color: "secondary",
                   },
                 }}
+                style={{ textDecoration: "none" }}
               >
-                {post.title}
-              </Styled.h2>
-            </Styled.a>
-            <Styled.p sx={{ fontSize: 1, color: "textGray", m: 0, mt: 2 }}>
-              {post.date} &bull;{" "}
-              <FaRegClock
-                sx={{
-                  position: "relative",
-                  top: "0.125em",
-                }}
-              />{" "}
-              {post.readingTimeInMinutes} Min
-            </Styled.p>
-            <Styled.p sx={{ mt: 2, mb: 0 }}>{post.excerpt}</Styled.p>
-            <Styled.a
-              as={Link}
-              to={rootPath.concat(post.slug.current.replace(/\/*$/, `/`))}
-              sx={{
-                transition: "color 0.2s ease",
-                ":hover, :focus, :active": {
-                  color: "secondary",
-                },
-              }}
-              style={{ textDecoration: "none" }}
-            >
-              Read More &rarr;
-            </Styled.a>
-          </article>
-        ))}
+                Read More &rarr;
+              </Styled.a>
+            </article>
+          ))}
+        </div>
       </Layout>
     </SanityThemeProvider>
   )