+++ themes/gatsby-theme-catalyst-bery/src/gatsby-theme-catalyst-sanity/components/queries/post-query.js
@@ -3,15 +3,22 @@ import { graphql } from "gatsby"
 import Post from "../../../components/post-template"
 
 export default ({ data }) => {
-  return <Post data={{ ...data }} />
+  const { previous, next } = data
+  return <Post data={{ ...data }} previous={previous} next={next} />
 }
 
 export const query = graphql`
-  query BeryPost($id: String!) {
+  query BeryPost($id: String!, $previousId: String, $nextId: String) {
     sanityPost(id: { eq: $id }) {
       id
       title
       date(formatString: "MMMM D, YYYY")
+      categories {
+        title
+        slug {
+          current
+        }
+      }
       readingTimeInMinutes
       featuredImage {
         caption
@@ -22,6 +29,24 @@ export const query = graphql`
         }
       }
       _rawBody
+    }
+    previous: sanityPost(id: { eq: $previousId }) {
+      id
+      excerpt
+      slug {
+        current
+      }
+      title
+      date(formatString: "MMMM DD, YYYY")
+    }
+    next: sanityPost(id: { eq: $nextId }) {
+      id
+      excerpt
+      slug {
+        current
+      }
+      title
+      date(formatString: "MMMM DD, YYYY")
     }
   }
 `